backend/requirements/base.txt
asgiref==3.7.2
Django==5.0
django-allauth==0.58.2
djangorestframework==3.14.0
dj-rest-auth==5.0.2
environs[django]
psycopg==3.1.13
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0